requirements.txt
flask
langchain
langchain_community
langchain_groq
python-dotenv
sentence_transformers
chromadb
pypdf
docx2txt